requirements.txt
certifi==2020.12.5
click==8.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
Werkzeug==2.0.2
yarl==1.7.2
slack==0.0.2
zipp==3.6.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
multidict==6.0.2
pyee==9.0.4
pyngrok==5.1.0
python-dotenv==0.19.2
PyYAML==6.0
slack-bolt==1.11.6
slack-sdk==3.15.2
slackclient==2.9.3
slackeventsapi==3.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
yarl==1.7.2